fix bugs in chart resizers
ciview-chart-titlebar and ciview-chart-content visibility set to hidden to stop
content overflowing

reisizer-xx z-index set to inherit to prevent problems when charts were overlapping

microtip (css tooltip) was replaced with tippy.js as the css tooltip was now obscured
due to hidden overflow in the title bar. Some logic had to be added to the chart's
changeBaseDocument in order to attach the tooltip to the correct document if the
chart was popped out.