Chart Resizer Fix (#261)
* updated chart resizer to move it outside the chart

* fix bugs in chart resizers

ciview-chart-titlebar and ciview-chart-content visibility set to hidden to stop
content overflowing

reisizer-xx z-index set to inherit to prevent problems when charts were overlapping

microtip (css tooltip) was replaced with tippy.js as the css tooltip was now obscured
due to hidden overflow in the title bar. Some logic had to be added to the chart's
changeBaseDocument in order to attach the tooltip to the correct document if the
chart was popped out.

* fix bugs in tippy.js

* added grid stack resizer styling and fixed tooltip issue in popout

* updated background to hide the duplicate resize handles at the bottom in grid stack

* updated comments for resize handler changes

* updated remove function to destroy menu tooltips

* reduced the height while destroying in grid stack to avoid extra black background at the bottom

* reverted z-index for the resizers

* added css variables for constant values for resizers

---------

Co-authored-by: martinSergeant <m.j.sergeant1@googlemail.com>